colorize: cast MB_CUR_MAX to size_t in mbtowc call
MB_CUR_MAX is correctly defined as size_t on Linux but is defined as an
int on macOS. Cast to size_t to silence -Wsign-conversion warning in the
mbtowc call.

Fixes: 1e5a3097f89c ("colorize: fix tab expansion to respect column position")
Acked-by: Robin Jarry <robin@jarry.cc>

+++ filters/colorize.c
@@ -473,7 +473,7 @@ static inline size_t print_notabs(const char *in, size_t max_len)
 		}
 
 		wchar_t wc;
-		int len = mbtowc(&wc, in, MB_CUR_MAX);
+		int len = mbtowc(&wc, in, (size_t)MB_CUR_MAX);
 
 		if (len < 1) {
 			/* Invalid sequence, output as-is and reset state */
